+++ include/triggerbot.hpp
@@ -6,7 +6,7 @@ using namespace hazedumper::netvars;
 using namespace hazedumper::signatures;
 
 extern class Memory Memory;
-extern uintptr_t client;
-extern uintptr_t player;
+extern uint32_t client;
+extern uint32_t player;
 
 void handleTriggerBot();
