+++ include/wallhack.hpp
@@ -22,9 +22,9 @@ struct ClrRender {
 };
 
 extern class Memory Memory;
-extern uintptr_t client;
-extern uintptr_t engine;
-extern uintptr_t player;
+extern uint32_t client;
+extern uint32_t engine;
+extern uint32_t player;
 
 void handleBrightness();
 void handleGlow();
